requirements/base.txt
gunicorn==19.2.1
Django==1.7.4
psycopg2==2.6
colorama==0.3.3
termcolor==1.1.0
dj-database-url==0.3.0
djangoajax==2.2.12
raven==5.2.0
djangorestframework==3.0.5
django-phonenumber-field==0.7.1
pillow==2.7.0
django-cors-headers==1.0.0
django-static-precompiler==0.9
django-registration-redux==1.1
django-taggit==0.12.2
django-widget-tweaks==1.3
djoser==0.2.0
django-polymorphic==0.6.1